+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/rerun-io/rerun/compare/latest...HEAD)
 * Add support for invisible tiles
 * `PartialEq` for `Tiles` now ignore internal state
+* Add `Tiles::find_pane`
 
 
 ## 0.1.0 - Initial Release - 2023-05-24